ITSM/requirements.txt
psycopg2-binary
dbt-core
apache-airflow
apache-superset
pandas
sqlalchemy